Rename sms module to messages
