+++ modules/main.js
@@ -9,7 +9,7 @@
     this.restore = require('./restore');
     this.clean = require('./clean');
     this.contacts = require('./contacts');
-    this.sms = require('./sms');
+    this.messages = require('./messages');
   };
 
   window.ffosbr = new Ffosbr();
